propagate vault path (#151)
With `space: enabled`, `vaultCommon` must be propagated from the operator
to the konks, konkservices, and example-apiservers. However, `overrideValues`
only supports strings. The important part of `vaultCommon` is the path,
so I've hardcoded the parts other than the path. We're left with a single
string (path) that needs to be propagated.